resume-analyzer/backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-jwt-extended==4.6.0
pymongo==4.7.3
bcrypt==4.1.3
python-dotenv==1.0.1
pdfplumber==0.11.0
python-docx==1.1.2
spacy==3.7.5
werkzeug==3.0.3
gunicorn==22.0.0
